Alguns debugs... e uns long long int (muito no desespero... hehe) Tá perdendo pacotes, botei um sleep lá, mas só ficou mais bizarro bom, fui nessa.
Abraço!

+++ include/data.h
@@ -17,7 +17,7 @@ struct data_info
 {
 	int name_size;			/**< Tamanho do nome do arquivo */
 	long int data_size;		/**< Tamanho do dado carregado */
-	long int tot_len;
+	long long int tot_len;
 	unsigned char fragmented;	/* 0 = No fragments, 1 = Is a fragment, 2 = Last Fragment */
 	int id;
 	int seq;
